# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Eden AI Workspace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pages:
    get:
      operationId: listPages
      tags: [pages]
      summary: List all pages in the workspace (flat tree)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Page"
    post:
      operationId: createPage
      tags: [pages]
      summary: Create a page
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
  /pages/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getPage
      tags: [pages]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageWithBlocks"
        "404":
          description: Not found
    patch:
      operationId: updatePage
      tags: [pages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePageInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
    delete:
      operationId: deletePage
      tags: [pages]
      responses:
        "204":
          description: Deleted
...
  /sources:
    get:
      operationId: listSources
      tags: [sources]
      summary: List text, URL, YouTube, image, and video sources
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Source"
    post:
      operationId: createSource
      tags: [sources]
      summary: Create a source from text, URL, YouTube, image, or video
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSourceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
  /sources/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getSource
      tags: [sources]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceWithChunks"
        "404":
          description: Not found
    patch:
      operationId: updateSource
      tags: [sources]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSourceInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
    delete:
      operationId: deleteSource
      tags: [sources]
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Page:
      type: object
      properties:
        id: { type: integer }
        kind:
          type: string
          enum: [page, folder]
        title: { type: string }
        emoji: { type: ["string", "null"] }
        parentId: { type: ["integer", "null"] }
        position: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, kind, title, emoji, parentId, position, createdAt, updatedAt]
    PageWithBlocks:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            blocks:
              type: array
              items:
                $ref: "#/components/schemas/Block"
          required: [blocks]
    CreatePageInput:
      type: object
      properties:
        kind:
          type: string
          enum: [page, folder]
        title: { type: string }
        emoji: { type: ["string", "null"] }
        parentId: { type: ["integer", "null"] }
      required: [title]
    UpdatePageInput:
      type: object
      properties:
        kind:
          type: string
          enum: [page, folder]
        title: { type: string }
        emoji: { type: ["string", "null"] }
        parentId: { type: ["integer", "null"] }
        position: { type: integer }
    Block:
      type: object
      properties:
        id: { type: integer }
        pageId: { type: integer }
        type:
          type: string
          enum: [text, heading1, heading2, heading3, todo, bulleted, numbered, quote, code, divider]
        content: { type: string }
        checked: { type: boolean }
        position: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, pageId, type, content, checked, position, createdAt, updatedAt]
...
    Source:
      type: object
      properties:
        id: { type: integer }
        kind:
          type: string
          enum: [text, url, youtube, image, video]
        title: { type: string }
        url: { type: ["string", "null"] }
        parentPageId: { type: ["integer", "null"] }
        mediaUrl: { type: ["string", "null"] }
        embedUrl: { type: ["string", "null"] }
        mediaMimeType: { type: ["string", "null"] }
        mediaSizeBytes: { type: ["integer", "null"] }
        summary: { type: ["string", "null"] }
        chunkCount: { type: integer }
        status:
          type: string
          enum: [processing, ready, error]
        createdAt: { type: string, format: date-time }
      required: [id, kind, title, url, parentPageId, mediaUrl, embedUrl, mediaMimeType, mediaSizeBytes, summary, chunkCount, status, createdAt]
    SourceWithChunks:
      allOf:
        - $ref: "#/components/schemas/Source"
        - type: object
          properties:
            content: { type: string }
            chunks:
              type: array
              items:
                $ref: "#/components/schemas/SourceChunk"
          required: [content, chunks]
    SourceChunk:
      type: object
      properties:
        id: { type: integer }
        sourceId: { type: integer }
        position: { type: integer }
        content: { type: string }
      required: [id, sourceId, position, content]
    CreateSourceInput:
      type: object
      properties:
        kind:
          type: string
          enum: [text, url, youtube, image, video]
        title: { type: string }
        content: { type: ["string", "null"] }
        url: { type: ["string", "null"] }
        parentPageId: { type: ["integer", "null"] }
        fileDataUrl: { type: ["string", "null"] }
        originalFilename: { type: ["string", "null"] }
        mediaMimeType: { type: ["string", "null"] }
      required: [kind, title]
    UpdateSourceInput:
      type: object
      properties:
        title: { type: string }
        parentPageId: { type: ["integer", "null"] }
...