# lib/api-zod/src/generated/types/sourceChunk.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * Eden AI Workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface SourceChunk {
  id: number;
  sourceId: number;
  position: number;
  content: string;
}